Dev-install packages in reverse dependency order
This is counter-intuitive, but seems to be the only way to work around
Pkg's behavior around dev-installs: Pkg always fully instantiates an
environment. In a project with depencies A, B where B depends on A, if
we dev-install A, Pkg will try to also install B (from the registry!).
This will fail if the *released* B is incompatible with the
dev-installed A.  However, if we dev-install B first, it will install a
released version of A, and then we can dev-install A to *overwrite* the
released version, and everything is OK